Move Tile components into their own folder
